Change TextMetrics.width to be Double

+++ src/main/scala/org/scalajs/dom/lib.scala
@@ -1208,7 +1208,7 @@ class TextMetrics extends js.Object {
    *
    * MDN
    */
-  var width: Int = ???
+  var width: Double = ???
 }
